Only install yarn assets if necessary
Yarn v2 (Berry) supports zero installs for frontend dependencies, so if
it's turned on, we don't need to install deps with yarn.

To determine if pnp/zero installs is enabled, we can check for the
existence of a `.pnp.js` file in the assets directory and install
dependencies the traditional way if it isn't.